+++ packages/components-js/tests/vrt/specs/marque.vrt.ts
@@ -32,8 +32,13 @@ it('should have no visual regression for :hover + :focus-visible', async () => {
   expect(
     await vrt.test('marque-states', async () => {
       const page = vrt.getPage();
-      const getElementsMarkup: GetMarkup = () => `<p-marque href="https://www.porsche.com"></p-marque>`;
-      await setContentWithDesignSystem(page, getBodyMarkup(getElementsMarkup));
+      const head = `<style>
+  p-marque:not(:last-child) { margin-right: 0.5rem; }
+</style>`;
+
+      const getElementsMarkup: GetMarkup = () => `<p-marque href="https://www.porsche.com"></p-marque>
+<p-marque href="https://www.porsche.com" style="padding: 1rem"></p-marque>`;
+      await setContentWithDesignSystem(page, getBodyMarkup(getElementsMarkup), { injectIntoHead: head });
 
       await forceFocusState(page, '.focus > p-marque'); // native outline should not be visible
       await forceFocusState(page, '.focus > p-marque >>> a');