+++ packages/components/src/components/form/text-field-wrapper/text-field-wrapper-utils.spec.ts
@@ -4,11 +4,11 @@ import {
   hasCounter,
   throwIfUnitLengthExceeded,
   TextFieldWrapperUnitPosition,
-  hasUnitAndIsTypeNumber,
   hasCounterAndIsTypeText,
   setCounterInnerHtml,
   addInputEventListener,
   setAriaElementInnerHtml,
+  hasUnitAndIsTypeTextOrNumber,
 } from './text-field-wrapper-utils';
 import * as textFieldWrapperUtils from './text-field-wrapper-utils';
 import { FormState } from '../../../types';
@@ -70,25 +70,25 @@ describe('hasCounterAndIsTypeText()', () => {
   );
 });
 
-describe('hasUnitAndIsTypeNumber()', () => {
-  it('should for input type="number" and unit="EUR" return true', () => {
+describe('hasUnitAndIsTypeTextOrNumber()', () => {
+  it.each<string>(['text', 'number'])('should for input type="%s" and unit="EUR" return true', (type) => {
     const inputElement = getInputElement();
-    inputElement.type = 'number';
-    expect(hasUnitAndIsTypeNumber(inputElement, 'EUR')).toBe(true);
+    inputElement.type = type;
+    expect(hasUnitAndIsTypeTextOrNumber(inputElement, 'EUR')).toBe(true);
   });
 
-  it('should for input type="number" and unit="" return false', () => {
+  it.each<string>(['text', 'number'])('should for input type="%s" and unit="" return false', (type) => {
     const inputElement = getInputElement();
-    inputElement.type = 'number';
-    expect(hasUnitAndIsTypeNumber(inputElement, '')).toBe(false);
+    inputElement.type = type;
+    expect(hasUnitAndIsTypeTextOrNumber(inputElement, '')).toBe(false);
   });
 
-  it.each<string>(['text', 'email', 'tel', 'search', 'url', 'date', 'time', 'month', 'week', 'password'])(
+  it.each<string>(['email', 'tel', 'search', 'url', 'date', 'time', 'month', 'week', 'password'])(
     'should for input type="%s" and unit="EUR" return false',
     (type) => {
       const inputElement = getInputElement();
       inputElement.type = type;
-      expect(hasUnitAndIsTypeNumber(inputElement, 'EUR')).toBe(false);
+      expect(hasUnitAndIsTypeTextOrNumber(inputElement, 'EUR')).toBe(false);
     }
   );
 });
@@ -178,7 +178,21 @@ describe('addInputEventListener()', () => {
     const ariaElement = getAriaElement();
     const spy = jest.spyOn(inputElement, 'addEventListener');
 
-    addInputEventListener(inputElement, counterElement, ariaElement);
+    addInputEventListener(inputElement, ariaElement, counterElement);
+    expect(spy).toHaveBeenCalledWith('input', expect.anything());
+  });
+
+  it('should register event listener on element without error when no counterElement is provided', () => {
+    const inputElement = getInputElement();
+    const ariaElement = getAriaElement();
+    const spy = jest.spyOn(inputElement, 'addEventListener');
+    let error = undefined;
+    try {
+      addInputEventListener(inputElement, ariaElement);
+    } catch (e) {
+      error = e;
+    }
+    expect(error).toBeUndefined();
     expect(spy).toHaveBeenCalledWith('input', expect.anything());
   });
 
@@ -189,7 +203,7 @@ describe('addInputEventListener()', () => {
 
     const setCounterInnerHtmlSpy = jest.spyOn(textFieldWrapperUtils, 'setCounterInnerHtml');
     const setAriaElementInnerHtmlSpy = jest.spyOn(textFieldWrapperUtils, 'setAriaElementInnerHtml');
-    addInputEventListener(inputElement, counterElement, ariaElement);
+    addInputEventListener(inputElement, ariaElement, counterElement);
 
     expect(setCounterInnerHtmlSpy).toHaveBeenCalledWith(inputElement, counterElement);
     expect(setCounterInnerHtmlSpy).toHaveBeenCalledTimes(1);
@@ -205,7 +219,7 @@ describe('addInputEventListener()', () => {
 
     const setCounterInnerHtmlSpy = jest.spyOn(textFieldWrapperUtils, 'setCounterInnerHtml');
     const setAriaElementInnerHtmlSpy = jest.spyOn(textFieldWrapperUtils, 'setAriaElementInnerHtml');
-    addInputEventListener(inputElement, counterElement, ariaElement);
+    addInputEventListener(inputElement, ariaElement, counterElement);
 
     inputElement.dispatchEvent(new Event('input'));
     expect(setCounterInnerHtmlSpy).toHaveBeenCalledWith(inputElement, counterElement);
@@ -220,7 +234,7 @@ describe('addInputEventListener()', () => {
     const counterElement = getCounterElement();
     const ariaElement = getAriaElement();
     const callback = jest.fn();
-    addInputEventListener(inputElement, counterElement, ariaElement, callback);
+    addInputEventListener(inputElement, ariaElement, counterElement, callback);
 
     inputElement.dispatchEvent(new Event('input'));
     expect(callback).toHaveBeenCalledTimes(1);