+++ packages/shared/rollup.config.js
@@ -4,10 +4,12 @@ import generatePackageJson from 'rollup-plugin-generate-package-json';
 import pkg from './package.json';
 
 const input = 'src/index.ts';
+const external = Object.keys(pkg.dependencies);
 
 export default [
   {
     input,
+    external,
     output: {
       dir: 'dist',
       format: 'cjs',
@@ -17,9 +19,6 @@ export default [
           baseContents: (packageJson) => ({
             ...packageJson,
             name: '@porsche-design-system/shared',
-            main: 'index.js',
-            module: 'esm/index.js',
-            types: 'types/index.d.ts',
             sideEffects: false,
             scripts: {},
             devDependencies: {},
@@ -36,6 +35,7 @@ export default [
   },
   {
     input,
+    external,
     output: {
       dir: 'dist/esm',
       format: 'esm',
@@ -45,7 +45,7 @@ export default [
   },
   {
     input: 'src/testing.ts',
-    external: Object.keys(pkg.dependencies),
+    external,
     output: {
       dir: 'dist',
       format: 'cjs',