+++ packages/utilities/projects/utilities/src/jss/typography/text/textXSmall.ts
@@ -1,0 +1,6 @@
+import { fontBehavior, fontFamily, fontStyle, fontVariant, fontWeight } from '../../font';
+
+export const textXSmall = {
+  font: `${fontStyle} ${fontVariant} ${fontWeight.regular} 0.75rem/1.6666666667 ${fontFamily}`,
+  ...fontBehavior,
+};