+++ packages/utilities/projects/utilities/src/jss/typography/textMedium.ts
@@ -1,6 +1,0 @@
-import { fontBehavior, fontFamily, fontStyle, fontVariant, fontWeight } from '../font/font';
-
-export const textMedium = {
-  font: `${fontStyle} ${fontVariant} ${fontWeight.regular} 1.5rem/1.5 ${fontFamily}`,
-  ...fontBehavior,
-};